tests/wpt/web-platform-tests/IndexedDB/idbobjectstore_get4.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBObjectStore.get() - attempt to retrieve a record that doesn't exist </title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>
    var db,
      t = async_test();

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;
        var rq = db.createObjectStore("store", { keyPath: "key" })
                   .get(1);
        rq.onsuccess = t.step_func(function(e) {
            assert_equals(e.target.results, undefined);
            setTimeout(function() { t.done(); }, 10);
        });
    };

    open_rq.onsuccess = function() {};
</script>

<div id="log"></div>
